[refactor] move ConsensusHandlerInitializer to consensus_handler file.
